+++ src/Observers/Observer.php
@@ -2,6 +2,6 @@
 
 namespace Sfneal\Observers;
 
-abstract class AbstractObserver
+abstract class Observer
 {
 }